break-out retrieve winning positions methods into their own class
